Move template missing option prompt to it's own file
