+++ src/cli.js
@@ -1,7 +1,6 @@
 import arg from 'arg';
-import inquirer from 'inquirer';
-import chalk from 'chalk';
 import { folderNameMissingOptionPrompt } from './prompts/folder';
+import { templateMissingOptionPrompt } from './prompts/template';
 import { createProject } from './main';
 
 let parseArgumentsIntoOptions = (rawArgs) => {
@@ -37,84 +36,6 @@ let parseArgumentsIntoOptions = (rawArgs) => {
     runInstall: args['--install'] || true,
     skipInstall: args['--skip-install'] || false
   }
-}
-
-let skipPrompts = (options, defaultFolderName, notAmongTemplateCollection, defaultTemplate) => {
-  if (notAmongTemplateCollection && options.template !== undefined) {
-    console.log( chalk.cyanBright(`Cli does not have template: "${options.template}" in its template collection, the default template: "${defaultTemplate}" will be used instead.`) );
-    options.template = defaultTemplate;
-  }
-
-  if (notAmongTemplateCollection && options.template === undefined) {
-    console.log( chalk.cyanBright(`--yes or -y flag detected. Generate "${defaultTemplate}" template since no template is specified.`) );
-    options.template = defaultTemplate;
-  }
-
-  options = {
-    ...options,
-    folderName: options.folderName || defaultFolderName,
-    template: options.template,
-    runInstall: false,
-    git: false
-  }
-
-  return options;
-}
-
-let templateMissingOptionPrompt = async (options, folderNameAnswers, defaultFolderName) => {
-  const defaultTemplate = 'esm';
-
-  const templateQuestions = [];
-
-  const templateCollection = [defaultTemplate, 'cjs', 'ts-esm'];
-
-  const equalToAtLeastOneTemplate = templateCollection.some(tc => {
-    return tc === options.template;
-  });
-
-  const notAmongTemplateCollection = equalToAtLeastOneTemplate === false;
-
-  if (!options.template || notAmongTemplateCollection) {
-    templateQuestions.push({
-      type: 'list',
-      name: 'template',
-      message: 'Please choose which project template to use',
-      choices: templateCollection,
-      default: defaultTemplate
-    });
-  }
-
-  let templateAnswers;
-
-  if (options.skipPrompts) {
-    options = skipPrompts(options, defaultFolderName, notAmongTemplateCollection, defaultTemplate);
-    templateQuestions.template = defaultTemplate;
-    templateAnswers = templateQuestions;
-  } 
-
-  if (notAmongTemplateCollection && options.template !== undefined && !options.skipPrompts) {
-    console.log( chalk.cyanBright(`Cli does not have template: "${options.template}" in its template collection`) );
-  }
-
-  if (!options.skipPrompts) templateAnswers = await inquirer.prompt(templateQuestions);
-
-  if (notAmongTemplateCollection) {
-    options = {
-      ...options,
-      folderName: options.folderName || folderNameAnswers.folderName,
-      template: templateAnswers.template,
-      git: options.git
-    }
-  }
-
-  options = {
-    ...options,
-    folderName: options.folderName || folderNameAnswers.folderName,
-    template: options.template || templateAnswers.template,
-    git: options.git
-  }
-
-  return options;
 }
 
 export let cli = async (args) => {
